Add support to install/uninstall/start a windows service
multipassd /install to install service and start it
multipassd /uninstall to stop and uninstall service
multipassd /svc will start the service and it's only meant for the
windows service manager to call.

multipassd with any other arguments can be called on the command line
so that it still starts standalone with the usual host of command line
options (for example, multipassd --logger=stderr), useful for
debugging and development.