cpack: windows: Fix PATH not being set
Fix PATH not being set correctly even when the user chooses to
modify it.

This only occurs when a previous install existed and its
uninstaller was run, which removed the PATH modifications in the
registry, but cannot modify the current environment string of the
process. Instead, read the latest environment directly from the
registry to figure out if the PATH entry we want to insert already
exists or not.