qemu: Remove QCoreApplication::processEvents() to avoid a crash
Due to how we are interacting with the Qt event loop, it's possible to crash
when deleting an instance while it's still launching. The after effect of this
is that concurrent client connections will block until launch (or other long
running commands) finishes.

A plan has been fleshed out to alleviate the concurrent blocking and will be
implemented later.

Fixes #585