Introduce Process, ProcessSpec and ProcessFactory
This is foundation work for adding security features to Multipass. The
goal is for processes that use external data to be confined with
technologies like AppArmor, and drop privileges if possible. This will
impact qemu, qemu-img and dnsmasq.

Note Multipass spawns many processes that we consider trusted, i.e
"ip" and "iptables" - these will remain unconfined.

This MP does not include any security mechanisms, just refactors
process management by multipass in preparation for it.

The design is the following:
- Process is a light wrapper around QProcess. It is intended solely for
  managing confined processes.
- ProcessSpec specifies the process and additional data relevant to the
  process. In future MP, this will include AppArmor rules, user to run as
  and other security tightening abilities.
- ProcessFactory takes a ProcessSpec, and returns a Process. In future MP,
  the process will have all the specified security mechanisms applied.