libvirt: Redo how the libvirtd connection is done and state machine
On any calls into the Libvirt backend, try a connection to libvirtd to ensure
it is there instead of keeping a presistent connection that can become stale.

Also, modify the whole state machine to better get the actual current state from
libvirt.

Modify and add new tests as seen fit for these modifications.