platform: Do not detect parent dying and send SIGHUP to ssfs_server
When starting mounts asynchronously, ie, on start, when the thread that starts the mount
exits, this code would send a SIGHUP to sshfs_process, causing it to exit prematurely.

This code is mostly unnecessary since it's only for Linux and systemd takes care of killing
child processes if the main process dies unexpectedly when running the snap.

Fixes #1143