[settings] Use UTF-8 for setting persistence
This makes the backend file human readable when there are "funny" chars,
but it does not affect logic, since the same format used is for reading
and writing. Furthermore, no transitioning is needed because no existing
settings accepted anything beyond ASCII, so they will be interpreted the
same way in UTF-8.