Switch to ssh_channel_request_pty_size to set TERM explicitly
Previously, all multipass sessions used the default TERM value
"xterm". With this change, we pull TERM in from the environment and
set it explicitly.

Also cleaned up the code that updates the pty size. This makes it
simpler to pull in geometry values for the initial call to request a
channel.