[test] Add a Scope inner struct to the MockLogger
Provide a MockLogger::Scope type that registers and unregisters the mock
logger, while providing access to the underlying mock. That makes it
more familiar, by following an approach similar to ProcessFactoryMock.

A MockLogger::Scope can only be created through MockLogger::inject.
This is a first step toward a MockLogger that can only be instantiated
this way, to prevent lingering mocks and repeated boilerplate.