Add option --network to launch daemon command.
This commit adds a list of network interfaces to the virtual machine
description. The first element of that list is the default interface, of
which only the mac address was recorded.

To avoid this change breaking compatibility with instances created with
prior versions of Multipass, the JSON reader accepts the old format,
containing only one MAC address.