Ask the factory how to match network interfaces.
All network interfaces marked as "auto" must be matched by cloud-init.
Most backends support setting the MAC address of the network interfaces,
what permits us to match them easily. However, if setting the MAC is not
possible, we need to match them by name. And the name depends on the
backend, so we need to ask the factory how to match.