Refactor VM creation.
The cloud-init file used to be written before the VM was created. This
had the disadvantage that we had to know how to match the network
interfaces beforehand. Now, the cloud-init file is created after the
VM was created. The backend is thus who tells us how to match each
interface, at VM creation time.