[daemon] Homogenize tactic for repeated-mac check
Replicate the strategy that the ctor already used to check mac addresses
for repetition, namely: work with a temporary copy of the allocated macs
and replace in the end if no problems are found. This simplifies the
code and decouples the the utility to generate unused macs from the
Daemon class, allowing it to work with an arbitrary set of used macs.
That may come in handy to check for mac repetition during db loading.