[daemon] Remove useless SSH session in list cmd.
In our first implementation, we used to create a SSH session in the
daemon to communicate with the instance to get the IP's it had in use.
We moved later creating the session to the VirtualMachine class, but we
forgot to remove the code in the daemon to create such session. As a
result, we created two sessions each time we wanted to ask an instance
for its IP's.