[daemon] Let create-vm exceptions through in ctor
Let exceptions that arise when creating a VM through. This avoids
erasing instance data that could be recoverable and important for the
user, but aborts the daemon since it can't honor its database, which
would otherwise be out of sync. More robust approaches for particular
causes can always added. Fixes #1658.