[daemon] Check first if the image exists.
When the requested image is in the list of the images which are not
capable of bridging, there is the possibility that it is not present on
the remote. It is better, in that case, to fail telling the user that
the image does not exist in the remote instead of saying that the image
is not bridgeable.