[daemon] Perform a health check on `networks`
Perform a hypervisor health check before inspecting networks. This is
required in the particular case of LXD because `mp::lxd_request` assumes
that the "multipass" project already exists, and that project is created
in the health check. More generally, it makes sense to require a healthy
backend before asking it for the network inspection. Fixes #2139.