[clang-format] Don't wrap braces before lambdas.
We used Allman-style brace formatting in clang-format 10. clang-format 12
incorporates a new behavior in the Allman style, which wraps the braces before
the lambdas, thus generating deeply indented lambda functions. We removed the
Allman style and set some options to their right values to mimic the behavior
we had in clang-format 10.