[clang-format] Fix indentation in lambdas.
This file was modified in commit 7524f1b221dd73ec0d5224230d0a82fe4b71c5b4.
clang-format wrapped the braces before the definition of the lambda. This
commit uses the new .clang-format file to fix it.

Note that this commit corrects only the wrong indentation of this specific
commit. It does not correct all the indentations of the file.