[todos] Update platform TODOs and reorder funcs
Now that only the daemon needs to know its config home, we should be
able to remove `daemon_config_home` in favor of a standard path.
However, this would impact multiple platforms + snap (and there is an
overriding env var involved). So drop that TODO, as it is better done
separately. Update remaining TODOs to nomerge scheme.