[client/common] Throw when unable to connect to daemon
In the logic to figure out if a cert has been accepted by the daemon, if the client cannot
establish a connection and get status if it's authenticated or not, it should just throw.

This will avoid the client from generating a new cert until communication with the daemon
is established.

Fixes #2552