Merge #2835
2835: Mount handlers refactor r=ricab,luis4a0,sharder996,townsend2010 a=andrei-toterman

This PR is a refactor of the mount handlers. Instead of having each mount handler keep track of all the mounts of a specific type, we will have each instance of a mount handler class represent a single mounted path. The daemon will keep track of these mounts and, since each mount can be thought of as a resource, we can use RAII to manage their creation and destruction to some extent. Previous tests have been removed in order to not block the next release, but they will be rewritten as soon as possible.



Co-authored-by: Andrei Toterman <andrei.toterman@canonical.com>