modify the list_instances function in the multipass-info.cpp and also added main  function to pass the show_deleted flag to it.
I've added a show_deleted parameter to control whether information about deleted instances should be included in the output. I've also added a check for deleted instances and a loop to print information about each deleted instance if the show_deleted flag is set.
And In order to use the modified list_instances function, we need to modify the info_main function to pass the show_deleted flag to it.