[shared] Use recursive mutexes for snapshots
We need the mutexes to be recursive, so that we can re-lock in the same
thread, as member functions can call each other and the Snapshot/VM
classes may call each other back. Unfortunately, there is no recursive
shared mutex in the standard library, nor could I find one in Qt or
POCO. I don't think we should bring another 3rd-party just for this use
case, so we loose the distinction between read/write access.