[image_host/common] Rework detecting if an image is supported via alias
The current way is too restrictive when checking against
platform-specific supported images being that *all* aliases must be
defined for a particular image in order for it to be supported.  This is
too restrictive and it should be that *any* alias that is defined should
make the image supported.

This is needed for proper output of `multipass find snapcraft:` on
macOS.