[ssh] Fix #3119.
When the user specified an empty string between quotes, some shells
(bash, in particular) passed the empty string without the quotes.
When this happened, the instance could not realize that there was
an empty string in the input; what leaded to errors. To fix this,
the empty strings are quoted again.

See https://github.com/canonical/multipass/issues/3119 for details.