[snapshots] Drop unsustainable consts
Make the Snapshot's pointer to its parent be non-const, adapting the
necessary cascade of non-constness from there on. Also change function
signatures such that parameters that can now be used for output come
last.

This allows a VM to update its head snapshot from another snapshot's
parent (e.g. to delete the current head), while retaining the power to
edit that snapshot (e.g. by applying it to revert a failed) restore. A
VM's head snapshot must be editable, e.g. so that it can be applied when
rolling back. Moreover, we need to be able to update the head snapshot
from another snapshot's  parent (e.g. to delete). In parallel, this
requires Snapshot constructors to receive non-const references to VMs,
which better reflects the fact that a snapshot can change the VM that it
pertains to.