[sftp] Reintroduce artificial condition for tests
Reintroduce an artificial condition that a complex arrangement of tests
depends on. Removing this would make many tests fail.

The existing `ExitStateMock` could be used to make `SftpServer` tests
succeed, at the cost of nearly doubling the time for a full test run
(due to an additional quarter-second in several tens of tests). However,
the `ExitStateMock` has no way to distinguish SSH processes. That would
be necessary for `SshfsMount` tests, since they rely on a number of SSH
commands beyond `sshfs`, which should complete right away. These tests
could of course all be rewritten with higher level mocks, but that would
be a very significant amount of work, far removed from present goals.

(Fixes broken tests.)