[factories] Remove obsolete method split
Remove obsolete separation between interface and implementation methods
for preparing networking at the level of the BaseVM. Provide the common
implementation directly in the base `prepare_networking` instead. This
needs to be overridden in backends that follow a different scheme (like
VirtualBox, where no preparation is needed). Adapt tests accordingly.