Update README.md
I've applied all the fixes suggested by the team (thanks Andrei, Jia and Scott!). Please review and see if it's all good now.

I have also been testing the instructions to Multipass, they need some debugging.

For now I'll leave the contributing info as a subsection of the Readme, but let's consider the possibility of having it as a separate file as it's done on other projects.

Signed-off-by: Giulia Zanchi <giulia.zanchi@canonical.com>