[exceptions] add FormattedExceptionBase type
FormattedExceptionBase is a generic base class type that produces
formatted string and passes it to the base class. This is a common
occurence in the codebase, so this class is designed to address that.

FormattedExceptionBase can be used with any type that derive from
std::exception and has a std::string constructor.

The class gracefully handle format errors in order to not incur
a `std::terminate()`.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>