[logging/cstring] introduce proper string-view log API
The CString type is unsafe and serves no purpose other than being a shim
for the real string's .c_str() function. CString class is a pseudo-string_view
class, but the logging API hasn't been designed in a way that could safely work
on string_views, and neither the CString class ensures that the input string is
NUL-terminated. The logging API and the CString class have the assumption that
the wrapped string will always be NUL terminated, which might not be the case
with the introduction of more string-view oriented API's.

This patch introduces a proper string-view based logging API that can accommodate
non-NUL terminated strings in a well-defined and safe way.

The new API is backward-compatible with the old one.

All the logging backends and the formatting utilities we use support the precision
controlled string format specifier %.*s, which allows us to express precisely how
many characters we want to print. That is exactly what was needed for properly
printing a std::string_view. The patch eliminates the CString type and uses
std::string_view instead while replacing the %s format specifiers with %.*s where
needed.

Summary of changes:

- Replaced the all instances of CString with std::string_view
- Removed the cstring.h header file
- Changed the templated log() functions signature to get category as std::string_view
  as we now can properly handle string_view's
- Replaced "%s" format specifier for sd_journal_send with "%.*s" and adjusted the format
  parameters accordingly.
- Replaced "%s" format specifier for syslog with "%.*s" and adjusted the format parameters
  accordingly.
- Removed the now-redundant make_cstring_matcher from all the tests

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>