[hyperv-api-vm] replace factory/vm hyperv_api_ pfx with hcs_
Hyper-V API has been the umbrella term to call the technology behind
this feature, but it might be misleading because there's actually a
real Hyper-V API and the feature does not use that. Instead, the feature
utilizes Host Compute System and Host Compute Network API's.

It's better to call this backend as what it actually is.