[ci/BuildAndTest] check for non-zero exit code instead of 139
sometimes the coverage target crashes, but even though the test executable returns
a 139, the build system swallows it and returns 2 instead. this is due to the fact
the covreport target being a CMake custom target, meaning the resulting target will
be run at build time, by the build tool.

the pull & upload steps now check for non-zero error code instead of segfault.
enabled continue-on-error for pull step so it won't cause job to fail when there's
no coredumps to full. it's already the case for the upload step (upload-artifact's
default behavior).

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>