[logging/linux] separate targets for journald/syslog for testing
the linux_logger's sources were conditionally being populated depending
on MULTIPASS_ENABLE_SYSLOG variable and availability of the journald.

The patch defines two separate targets for journald and syslog, which
would be defined when the journald and syslog are available in the environment
respectively.

The linux_logger is now an alias target depending on MULTIPASS_ENABLE_SYSLOG's
value.

The journald/syslog tests now run

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>