[hyperv-api-factory] implement remove_resources_for_impl
There are nothing to remove since everything is neatly packed into
the instance folder. Removal of the instance folder is already taken
care of by the base class, so only thing left is to ensure that the
compute system is terminated.

Set log level of open_compute_system failure and operation_failure
to debug() instead of error() since these functions may fail in normal
operation workflow (i.e. check if a VM exists).