[common_callbacks] fix make_confirmation_callback lifetime issue
The make_confirmation_callback function takes the key argument as
const ref, meaning that it'll be bound to an existing QString when
the arg is QString, but a temporary QString would be constructed and
bound to it when the "key" is not a QString. This is fine during the
make_confirmation_callback's scope since const& extend the lifetime
of the temporary, but passing that "const&" to a capture list of a lambda
that'll be executed later does not extend the lifetime, hence leading
to undefined behavior.

The patch fixes this issue by taking QString as a value and moving that
into the lambda capture.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>