[3rd-party/CMakeLists] Define CMAKE_POLICY_VERSION_MINIMUM=3.5
With CMake 4.0 support for compatibility with versions older than
3.5 has been removed. Some of our dependencies (yaml-cpp, semver) define
their policy versions as <=3.4, so the build with CMake 4 fails for that reason.

There's an option to override that behavior, called CMAKE_POLICY_VERSION_MINIMUM.
This flag allows CMake to override a subdirectory's cmake_minimum_required() call
without having to modify it.

note: yaml-cpp builds without any issues when option is in place. The upstream
also updated the cmake_minimum_required() to `3.4...3.14` without making any code
changes so it's safe to assume that project is compatible with the CMake 3.5 policies.
it's the same for the semver as well.

https://github.com/jbeder/yaml-cpp/commit/c2680200486572baf8221ba052ef50b58ecd816e
https://github.com/Neargye/semver/commit/df774773960913c5045cb6d8d7720fa5aa2f92dc

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>