[cloud-init/netplan] explicitly name interfaces with `set-name`
until now, the interface names were assigned automatically. the names
were given in a "first-come-first-served" fashion, i.e. whichever interface
initializes first gets the "eth0", regardless of the declaration order.
this makes the interface names inconsistent, which makes it harder to
identify the "extra" interfaces.

this patch explicitly names the default interface as "primary" and the
extra ones as "extra<N>" where the <N> is the zero-based index for the
extra interface.

also tidied up the yaml_node_utils a bit.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>