[hyperv-virtdisk-snapshot] implement the snapshot feature
The snapshot feature is implemented based on virtdisk's differencing
disks. When the user wants to create a new snapshot, the implementation
creates a new differencing disk for the snapshot. The created snapshot
disk is read-only. Another disk called "head" is created for the future
writes. When the user wants to create another snapshot, the head becomes
the new snapshot and a new empty "head" is created.

When user wants to restore to a previous snapshot, the head moves from
one parent to the another. The contents in head won't be preserved as
the block histories between the parents are not compatible. The user
has option to preserve the existing head's state by creating a new
snapshot.

When a snapshot is deleted, its' contents are merged into the parent
and all the children disks are reparented to the parent.

- open_virtual_disk now takes access_mask, flags and params as parameter
- added merge_virtual_disk_to_parent function
- added reparent_virtual_disk function