[ci] add vcpkg binary caching (#4155)
This PR introduces [binary
caching](https://learn.microsoft.com/en-us/vcpkg/consume/binary-caching-github-packages?pivots=linux-runner)
for vcpkg. Since going open source, we can take advantage of [GitHub
Packages](https://docs.github.com/en/packages/learn-github-packages/introduction-to-github-packages)
to host binary caches.

Binary caching is enabled by setting the `VCPKG_BINARY_SOURCES`
environment variable. A tutorial for how to do this can be found
[here](https://learn.microsoft.com/en-us/vcpkg/consume/binary-caching-github-packages)
and was generally followed for changes in this PR.

For MacOS and ARM, there is seemingly no official support or guide, and
with the differences in bundled packages in GitHub hosted runners, a
little bit of tailoring is required. On non-Windows machines, once a
binary source has been registered, vcpkg seems to be hard coded to call
`mono nuget` in order to restore or upload binaries. `mono` expects a
.Net assembly compiled for Intel architecture, so when an ARM compatible
package is bundled in the image from GitHub this fails. To get around
this, NuGet is manually downloaded and symlinked so that vcpkg will call
it.

On Linux, binary caching must be configured within the snapcraft image.
Again the appropriate NuGet binary needs to be downloaded and configured
so that it is callable from the command line with `mono nuget`.

Build times are considerably shorter once all dependencies are cached
with the "Configure" step taking less than a few minutes. On average,
this saves around 30 minutes which is approximately a 33-50% improvement
in build times.

Because of the expectancy that self-hosted runners will be dropped soon,
binary caching is disabled for them.

GitHub Packages does not offer any sort of retention policy meaning that
as vcpkg binaries change versions, Packages will increasingly become
more and more cluttered with old versions of binaries. A solution to
this problem is to setup a scheduled GitHub Action with
[actions/delete-package-versions](https://github.com/marketplace/actions/delete-package-versions)
and `min-versions-to-keep: 1` to only keep the most recent version of a
binary.

There are two storage plans for GitHub Packages which depends on
visibility. Publicly visible packages show up on the [front
page](https://github.com/orgs/canonical/packages?repo_name=multipass) of
the Multipass repo, but have unlimited storage and bandwidth quota.
Having these packages there is a bit unsightly and to the uneducated
visitors they are seemingly irrelevant to Multipass and might confuse
someone trying to download the Multipass application. Setting the
visibility to private would be preferential, but is subject to
[billing](https://docs.github.com/en/billing/managing-billing-for-your-products/about-billing-for-github-packages).

An alternative solution for storage is to use our [AWS S3
bucket](https://learn.microsoft.com/en-us/vcpkg/reference/binarycaching#aws).
At this point, it is unclear whether this removes binaries from GitHub
Packages or is simply a different backend storage provider for GitHub
Packages. Also, since I do not have access to Multipass' AWS account,
available storage quotas and unknown and so whether this is a viable
solution is also unknown.

---

P.S. You can even setup binary caching
[locally](https://learn.microsoft.com/en-us/vcpkg/consume/binary-caching-local?pivots=shell-bash)
to speed up builds!

MULTI-2030