Fix client key permissions (#4194)
There was a change in logic in when permissions are set for ssl
certificates which caused newly created client certificates to have
overly restrictive permissions. This fixes that.

Also there was a bug where unix permissions weren't converted to Windows
permissions correctly. This fixes that.

Setting the permissions of the file to read by everyone does not change
security since the parent folder maintains the user only access.
However, in the future it would be nice to give read permissions
directly to the user instead of everyone.

closes #4193