[client/gui] symlink libdart_ffi to bundle/lib post build
The GUI binary, multipass.gui's RUNPATH is $ORIGIN/lib, which resolves
to `$REPO_ROOT/build/bin/bundle/lib/`, whereas the libdart_ffi.so is in
`$REPO_ROOT/build/lib` folder. Hence, the multipass.gui fails to launch
on local builds if libdart_ffi.so is not somehow placed to the library
search path manually (i.e. by copying, or by LD_LIBRARY_PATH).

This patch adds a libdart_ffi.so symlink to the bundle/lib so the GUI
can find the dart_ffi library on local build without relying on a manual
intervention.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>