[cmake] Fix version derivation on win/mac
On Windows and macOS, our version includes a "win" or "mac" suffix in
the <build> component. In actual releases, that is all that the <build>
component consists of. This updates that part of the version derivation
logic such that the semantic version grammar is respected and WiX
accepts our version. We no longer have "+full" in our release tags and
we need our suffixes to be properly separated from the rest of the
version. In particular, they need to come after a single plus ('+') and
must be separated from other parts of the <build> component with dots
('.').

See
https://semver.org/#backusnaur-form-grammar-for-valid-semver-versions.