[git] Give up checking rule 9
Give up checking whether rule 9 holds because there is no way to know
with certainty whether a sentence-ending line also ends the paragraph,
without relying on a blank line to separate it from the following one.
As, for example, in the previous line and this one :)

We could detect cases where the first word(s) of the following line
could still fit in the line that ended the sentence, but that could also
be just a case of poor wrapping. For example:

This is a sentence.
And this is another one.

There is no way to decide if the above is a case of a paragraph without
a blank line or just sub-optimal wrapping. I am opting not to enforce
optimal wrapping at this point, since it is not explicitly required by
the current rules. And poems should be allowed :)