[ssl-cert-provider] regenerate gRPC server certs on cert mismatch (#4226)
The previously generated gRPC server certificate's issuer could be
different than the current one in certain circumstances, like a manual
installation after a snap installation of Multipass.

The patch fixes that by loading both the root and the subordinate cert
for the gRPC server and verifying that the root certificate is the
issuer.

Also added some logs to the certificate regeneration logic.

Fixes #4218

MULTI-2100