[sshfs-mount] do not call chown when id is no_id_info_available
chown is a platform-specific function which depends uid and gid.
these values are not available in Windows, and the chown impl
is by default returns -1. This causes filesystem operation failures
in Windows hosts when sshfs mounts are used in guests. One example
is as follows:

ubuntu@foo:~/Downloads$ echo "test" > foo.txt
-bash: foo.txt: Operation not permitted

This operation fails because it first lands in handle_open, which
eventually calls chown to change the file's owner, which is a failure
by default in Windows.

The patch fixes that by introducing a shim function, chown_shim which
returns success when either uid or gid is set to no_id_info_available.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>