[tests/cloud_init_iso] fix stack overrun by read mock
read_returns_five_bytes_string mock was using std::strcpy to copy
five characters from the string "NonJo". In reality, strcpy always
terminates the destination with a '\0', which leads to an overflow
since the destination is a 5-byte stack-allocated buffer.

This patch fixes that by replacing std::strcpy with std::memcpy.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>