[cli-tests] implement a pywinpty based pexpect-spawn class
pexpect's PopenSpawn is having a hard time interacting with Linux
terminals since they expect a pty/tty to be available:

bash: cannot set terminal process group (-1): Inappropriate ioctl for device
bash: no job control in this shell

I've tried to trick the terminal believing as if a PTY is available,
but still the interaction was clunky due to pipe-based behavior.
The outputs from commands like "echo" and "pwd" simply were absent
in the stdout.

To overcome this, a pty is needed. WinptySpawn class inherits from
pexpect.SpawnBase class just like the PopenSpawn and implements the
same functionality -- the only difference being it uses PtyProcess
instead of subprocess.