[cli-tests] introduce abstract daemon controller
The current async controller does two jobs at once, which are to interact
with the daemon, and control the daemon's lifecycle. This patch abstracts
the first part as "controller". The remanining part is now named as
"governor". Governor now uses the abstract controller interface to interact
with the daemon, which allows us to implement different ways of interacting
with the daemon in different deployment types.

The current available daemon controller, which is running the daemon directly
from the bin/ folder is named as the "standalone" controller. The other deployment
types (e.g. snap, launchd, winsvc) will have their respective controller type
to make the tests work with different deployments.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>