[linux/apparmor] implement fork-safe aa_change_onexec alternative
aa_change_onexec is designed to be run before an exec(), but not necessarily
a fork() before in mind. a fork()'ed process is often in a special state that
only a handful of functions can be called and the memory allocation opportunities
are also limited. The new aa_change_onexec_forksafe function mimics aa_change_onexec
using only the allowed functions and no memory allocation.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>