[test-platform-linux] Override libvirt.so properly
The with_minimally_mocked_libvirt function was constructing a new
libvirt wrapper before the test body, but it's essentially a no-op
since the libvirt factory uses its own instance of the wrapper.

The patch fixes that by properly overriding the default libvirt.so
path so the mock implementation would be loaded.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>