[backend-utils] Validate IP before returning
Ubuntu Core images initialize themselves in two stages separated by a
guest system reboot. An IP address is assigned to the guest during the
first phase even though networking is not fully initialized. We wait
until a fresh IP is assigned and the guest signals it is ready by
sending a response from a ping.