[cmake] Remove deprecated decl from libssh build
Most of the existing warnings and errors coming from the deprecated
function declarations from the libssh library come from the library
itself. It includes the legacy.h header in its main header libssh.h and
that header contains all the deprecated functions. Since it has to be
built and the warnings come from internal use in deprecated functions of
other deprecated functions, the most pragmatic and correct solution is
to silence the warnings coming from building libssh itself.

The usage of deprecated functions outside libssh will still raise a
warning or error.