[daemon] Improve file URL error message (#4198) (#4352)
Local file URLs should have three slashes after "file:".

This just adds an error message alerting the user to the fact that
non-local file URLs aren't supported (at least, not yet), and suggesting
a fix. The wording feels a little clumsy to me, but I'm not sure of a
better way to phrase it without becoming confusing.

(Recreated this from #4341 in the original repo so CI passes.)