[sftp] Replace sftp_server_init
The function is replace with its existing internal functionality.
Because in this case we cannot rely on the abstraction of S
SSH:throw_on_error due to the libssh functions not setting the ssh_error
on failure (previously the role of the replaced function), the exception
management is done directly.